backend/requirements.txt
Django==4.2.16
psycopg2-binary==2.9.9
redis==5.0.8
django-cors-headers==4.4.0
gunicorn==22.0.0
python-dotenv==1.0.1
PyJWT==2.8.0
